Notes about the client
