+++ worlds/apquest/__init__.py
@@ -1,20 +1,5 @@
-from worlds.LauncherComponents import Component, Type, components, launch_subprocess
-
+# In our __init__.py, we just import our world class from our world.py to initialize it.
+# Obviously, this world class needs to exist first. For this, read world.py.
+# Client stuff, don't worry about this.
+from .client.launch import run_client as run_client
 from .world import APQuestWorld as APQuestWorld
-
-
-def run_client(*args) -> None:
-    from .client.launch import launch_client  # lazy import
-
-    launch_subprocess(launch_client, name="APQuest Client", args=args)
-
-
-components.append(
-    Component(
-        "APQuest Client",
-        func=run_client,
-        game_name="APQuest",
-        component_type=Type.CLIENT,
-        supports_uri=True,
-    )
-)
