+++ worlds/apquest/client/ap_quest_client.py
@@ -20,6 +20,13 @@ if TYPE_CHECKING:
     import kvui
 
 
+# !!! IMPORTANT !!!
+# The client implementation is *not* meant for teaching.
+# Obviously, it is written to the best of its author's abilities,
+# but it is not to the same standard as the rest of the apworld.
+# Copy things from here at your own risk.
+
+
 class ConnectionStatus(Enum):
     NOT_CONNECTED = 0
     SCOUTS_NOT_SENT = 1
